+++ jwst/pipeline/calwebb_detector1.py
@@ -20,6 +20,7 @@ from ..dark_current import dark_current_step
 from ..reset import reset_step
 from ..persistence import persistence_step
 from ..jump import jump_step
+from ..undersampling_correction import  undersampling_correction_step
 from ..ramp_fitting import ramp_fit_step
 from ..gain_scale import gain_scale_step
 
@@ -60,6 +61,7 @@ class Detector1Pipeline(Pipeline):
                  'reset': reset_step.ResetStep,
                  'persistence': persistence_step.PersistenceStep,
                  'jump': jump_step.JumpStep,
+                 'undersampling_correction': undersampling_correction_step.UndersamplingCorrectionStep,
                  'ramp_fit': ramp_fit_step.RampFitStep,
                  'gain_scale': gain_scale_step.GainScaleStep,
                  }